This is a new commit made with an editor!
This is the commit message that will be used for the commit. The commit message is the text that will be associated with the commit. It is important to write a clear and concise message that describes the changes made in the commit. The commit message should be written in the present tense and should be no longer than 50 characters. If the commit message is longer than 50 characters, it will be truncated when displayed in the commit history.

Once you have finished writing the commit message, save and close the editor. The commit will be created and the changes will be saved to the repository.

End message.